move CLI to its own module
