+++ CLI/README.md
@@ -1,0 +1,8 @@
+# GaPLAC Command Line Interface
+
+## Usage
+
+
+
+## Examples
+
