+++ Project.toml
@@ -4,24 +4,19 @@ version = "0.1.0"
 
 [deps]
 AbstractGPs = "99985d1d-32ba-4be9-9821-2ec096f28918"
-AbstractGPsMakie = "7834405d-1089-4985-bd30-732a30b92057"
-ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 GPLikelihoods = "6031954c-0455-49d7-b3b9-3e1c99afaf40"
 KernelFunctions = "ec8451be-7e33-11e9-00cf-bbf324bd1392"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-LoggingExtras = "e6f89c97-d47a-5376-807f-9c37f3926c36"
 ReTest = "e0db7c4e-2690-44b9-bad6-7687da720f89"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 
 [compat]
-ArgParse = "1"
 CSV = "0.8"
 DataFrames = "1"
 Distributions = "0.21 - 0.25"
