+++ bin/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+CLI = "3c4b6d91-1f5f-4463-b11d-7f4de956452e"
+GaPLAC = "5fbfd2c5-e7c0-480c-b0aa-53d89672c7de"
