--- bin/cli/GaPLAC_CLI.jl
@@ -1,7 +1,0 @@
-module CLI
-
-include("mcmc.jl")
-include("sample.jl")
-include("select.jl")
-
-end # module
