+++ gaplac
@@ -1,3 +1,3 @@
 #!/usr/bin/sh
-
-julia --startup-file=no -t auto --project=@. --color=yes ./bin/main.jl "$@"
+BASEDIR="$(dirname $0)/bin"
+julia --startup-file=no -t auto --project="$BASEDIR" --color=yes "$BASEDIR/main.jl" "$@"
